KnowledgeGraph-TabularData/requirements.txt
gradio
ipykernel
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langsmith
numpy
openai
pandas
pillow
pyprojroot
PyYAML
SQLAlchemy
